+++ packages/yourloops/test/integration/pages/home-page/caregiver-home-page.spec.tsx
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright (c) 2022, Diabeloop
  *
  * All rights reserved.
@@ -27,21 +27,23 @@
 
 import PatientAPI from '../../../../lib/patient/patient-api'
 import { checkSecondaryBar } from '../../utils/patientSecondaryBar'
-import { mockAuth0Hook } from '../../mock/mockAuth0Hook'
+import { loggedInUserId, mockAuth0Hook } from '../../mock/mockAuth0Hook'
 import { mockNotificationAPI } from '../../mock/mockNotificationAPI'
-import { mockDirectShareApi } from '../../mock/mockDirectShareAPI'
-import { buildPatient, mockPatientAPI } from '../../mock/mockPatientAPI'
+import { mockDirectShareApi, removeDirectShareMock } from '../../mock/mockDirectShareAPI'
+import { buildPatient, mockPatientAPI, monitoredPatient, unmonitoredPatient } from '../../mock/mockPatientAPI'
 import { mockUserDataFetch } from '../../mock/auth'
 import { mockTeamAPI } from '../../mock/mockTeamAPI'
 import { checkCaregiverLayout } from '../../assert/layout'
 import { UserRoles } from '../../../../models/user'
 import { renderPage } from '../../utils/render'
-import { screen, within } from '@testing-library/react'
+import { act, screen, within } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
+import DirectShareApi from '../../../../lib/share/direct-share-api'
 
 describe('Caregiver home page', () => {
   const firstName = 'Eric'
   const lastName = 'Ard'
+
   beforeAll(() => {
     mockAuth0Hook(UserRoles.caregiver)
     mockNotificationAPI()
@@ -62,6 +64,7 @@ describe('Caregiver home page', () => {
     const patient1 = buildPatient({
       userId: 'patientId1',
       profile: {
+        email: 'Akim@embett.com',
         firstName: 'Akim',
         lastName: 'Embett',
         fullName: 'Akim Embett',
@@ -71,6 +74,7 @@ describe('Caregiver home page', () => {
     const patient2 = buildPatient({
       userId: 'patientId2',
       profile: {
+        email: 'alain@provist.com',
         firstName: 'Alain',
         lastName: 'Provist',
         fullName: 'Alain Provist',
@@ -80,6 +84,7 @@ describe('Caregiver home page', () => {
     const patient3 = buildPatient({
       userId: 'patientId3',
       profile: {
+        email: 'annie@versaire.com',
         firstName: 'Annie',
         lastName: 'Versaire',
         fullName: 'Annie Versaire',
@@ -119,5 +124,88 @@ describe('Caregiver home page', () => {
     expect(patientTableBody.getByText(patient2.profile.fullName)).toBeVisible()
     expect(patientTableBody.queryByText(patient1.profile.fullName)).not.toBeInTheDocument()
     expect(patientTableBody.queryByText(patient3.profile.fullName)).not.toBeInTheDocument()
+  })
+
+  it('should display a list of patients and allow to remove one of them', async () => {
+    const patientFullName = 'Unmonitored Patient'
+
+    await act(async () => {
+      renderPage('/')
+    })
+
+    checkCaregiverLayout(`${firstName} ${lastName}`)
+    checkSecondaryBar(false, false)
+
+    const patientTableBody = within(screen.getByTestId('patient-table-body'))
+    const patientData = patientTableBody.getByText(patientFullName)
+    expect(patientData).toBeVisible()
+
+    const patientRow = screen.queryByTestId(`patient-row-${unmonitoredPatient.userId}`)
+    const removePatientButton = within(patientRow).getByRole('button', { name: 'Remove patient-ylp.ui.test.patient28@diabeloop.fr' })
+    expect(removePatientButton).toBeVisible()
+
+    userEvent.click(removePatientButton)
+
+    const removePatientDialog = screen.getByRole('dialog')
+    expect(removePatientDialog).toBeVisible()
+
+    const removePatientDialogTitle = within(removePatientDialog).getByText('Remove a patient')
+    expect(removePatientDialogTitle).toBeVisible()
+
+    const removePatientDialogQuestion = within(removePatientDialog).getByText('Are you sure you want to remove patient Unmonitored Patient?')
+    expect(removePatientDialogQuestion).toBeVisible()
+
+    const removePatientDialogCancelButton = within(removePatientDialog).getByText('Cancel')
+    expect(removePatientDialogCancelButton).toBeVisible()
+
+    const removePatientDialogConfirmButton = within(removePatientDialog).getByRole('button', { name: 'Remove patient' })
+    expect(removePatientDialogConfirmButton).toBeVisible()
+
+    userEvent.click(removePatientDialogCancelButton)
+
+    expect(removePatientDialog).not.toBeInTheDocument()
+
+    userEvent.click(removePatientButton)
+
+    const removePatientDialog2 = screen.getByRole('dialog')
+    expect(removePatientDialog2).toBeVisible()
+
+    const removePatientDialog2ConfirmButton = within(removePatientDialog2).getByRole('button', { name: 'Remove patient' })
+
+    jest.spyOn(PatientAPI, 'getPatients').mockResolvedValueOnce([monitoredPatient])
+    await act(async () => {
+      userEvent.click(removePatientDialog2ConfirmButton)
+    })
+
+    expect(removeDirectShareMock).toHaveBeenCalledWith(unmonitoredPatient.userId, loggedInUserId)
+    expect(patientData).not.toBeInTheDocument()
+    expect(screen.queryByTestId('remove-direct-share-dialog')).toBeFalsy()
+    expect(screen.getByTestId('alert-snackbar')).toHaveTextContent('You no longer have access to your patient\'s data.')
+  })
+
+  it('should display an error message if patient removal failed', async () => {
+    jest.spyOn(DirectShareApi, 'removeDirectShare').mockRejectedValueOnce('Error')
+
+    await act(async () => {
+      renderPage('/')
+    })
+
+    const patientRow = screen.queryByTestId(`patient-row-${unmonitoredPatient.userId}`)
+
+    const removeButton = within(patientRow).getByRole('button', { name: `Remove patient-${unmonitoredPatient.email}` })
+
+    userEvent.click(removeButton)
+
+    const removeDialog = screen.getByRole('dialog')
+
+    const confirmRemoveButton = within(removeDialog).getByRole('button', { name: 'Remove patient' })
+
+    await act(async () => {
+      userEvent.click(confirmRemoveButton)
+    })
+
+    expect(removeDirectShareMock).toHaveBeenCalledWith(unmonitoredPatient.userId, loggedInUserId)
+    expect(screen.getByTestId('remove-direct-share-dialog')).toBeVisible()
+    expect(screen.getByTestId('alert-snackbar')).toHaveTextContent('Impossible to remove patient. Please try again later.')
   })
 })